LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-conditionals-expected.txt


FAIL Basic @supports assert_equals: expected "PASS" but got "--f()"
FAIL Basic @supports (false) assert_equals: expected "PASS" but got "--f()"
FAIL Nested @supports assert_equals: expected "PASS" but got "--f()"
FAIL Nested @supports (false) assert_equals: expected "PASS" but got "--f()"
FAIL Inconsequential conditional assert_equals: expected "PASS" but got "--f()"
FAIL @supports with locals assert_equals: expected "1 20 3" but got "--f(1)"
FAIL Basic @media assert_equals: expected "PASS" but got "--f()"
FAIL Basic @media (false) assert_equals: expected "PASS" but got "--f()"
FAIL Nested @media assert_equals: expected "PASS" but got "--f()"
FAIL Nested @media (false) assert_equals: expected "PASS" but got "--f()"
FAIL Locals within @media assert_equals: expected "PASS" but got "--f()"
FAIL @supports within @media assert_equals: expected "PASS" but got "--f()"
FAIL @media within @supports assert_equals: expected "PASS" but got "--f()"
FAIL Basic @container assert_equals: expected "PASS" but got "--f()"
FAIL Basic @container (false) assert_equals: expected "PASS" but got "--f()"
FAIL Nested @container assert_equals: expected "PASS" but got "--f()"
FAIL Nested @container (false) assert_equals: expected "PASS" but got "--f()"
FAIL Locals within @container assert_equals: expected "PASS" but got "--f()"
FAIL @supports within @container assert_equals: expected "PASS" but got "--f()"
FAIL @container within @supports assert_equals: expected "PASS" but got "--f()"
FAIL @container, @media, @supports assert_equals: expected "PASS" but got "--f()"
FAIL @supports, @media, @container assert_equals: expected "PASS" but got "--f()"
